requirements.txt
openai-whisper
pydub
ffmpeg-python
torch
